+++ src/domains/accounts/profiles/profiles.resource.js
@@ -1,0 +1,1 @@
+// this file will hold resource methods for profile.
